+++ docs/example.tf
@@ -7,4 +7,17 @@ module "example_project" {
   github_repo_branch            = "main"
   buildspec                     = "buildspec.yml"
   github_repo_url               = "https://github.com/example/example-project.git"
+
+  environment_variables = [
+    {
+      name  = "SOME_VARIABLE",
+      value = "some-value",
+      type  = "PLAINTEXT"
+    },
+    {
+      name  = "SOME_OTHER_VARIABLE",
+      value = "some-other-value",
+      type  = "PLAINTEXT"
+    }
+  ]
 }