+++ variables.tf
@@ -80,3 +80,26 @@ variable "vpc_security_group_ids" {
   default     = [""]
   description = "The IDs of the security groups for the CodeBuild project."
 }
+
+variable "permissions_boundary" {
+  type        = string
+  description = "ARN of the policy that is used to set the permissions boundary for the role."
+  default     = ""
+}
+
+variable "environment_variables" {
+  type = list(object({
+    name  = string,
+    value = string,
+    type  = string
+  }))
+  description = "The environment variables to create for the CodeBuild project."
+  default     = []
+
+  validation {
+    condition = alltrue([
+      for obj in var.environment_variables : contains(["PARAMETER_STORE", "PLAINTEXT", "SECRETS_MANAGER"], obj.type)
+    ])
+    error_message = "Type can only be PARAMETER_STORE, PLAINTEXT, or SECRETS_MANAGER."
+  }
+}